+++ database/migrations/2020_02_26_173434_create_category_product_table.php
@@ -15,9 +15,9 @@ class CreateCategoryProductTable extends Migration
     {
         Schema::create('category_product', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('category_id')->unsigned();
-            $table->integer('product_id')->unsigned();
-      
+            $table->bigInteger('category_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
+
             $table->foreign('category_id')->references('id')->on('categories');
             $table->foreign('product_id')->references('id')->on('products');
             $table->timestamps();
